+++ StarshipArena/src/Transport.java
@@ -14,15 +14,15 @@ public class Transport extends Starship{
 	static Texture red_tex3 = new Texture("red_transport3.png");
 	static Texture red_tex4 = new Texture("red_transport4.png");
 	
-	public Transport(StarshipArena mygame, int spawnx, int spawny){
+	public Transport(StarshipArena mygame, double spawnx, double spawny){
 		super(mygame, spawnx, spawny);
 	}
 	
-	public Transport(StarshipArena mygame, int spawnx, int spawny, int spawnangle){
+	public Transport(StarshipArena mygame, double spawnx, double spawny, double spawnangle){
 		super(mygame, "none", spawnx, spawny, spawnangle);
 	}
 
-	public Transport(StarshipArena mygame, String newteam, int spawnx, int spawny, int spawnangle){
+	public Transport(StarshipArena mygame, String newteam, double spawnx, double spawny, double spawnangle){
 		super(mygame, newteam, spawnx, spawny, spawnangle);
 	}
